refactor(models): polish schemas, comments and code locations
