+++ packages/cli/src/lib/implementation/base-command-config.ts
@@ -1,11 +1,11 @@
 import {
-  globalCliArgsSchema,
+  globalOptionsSchema,
   refineCoreConfig,
   unrefinedCoreConfigSchema,
 } from '@quality-metrics/models';
 import { z } from 'zod';
 
 export const commandBaseSchema = refineCoreConfig(
-  globalCliArgsSchema.merge(unrefinedCoreConfigSchema),
+  globalOptionsSchema.merge(unrefinedCoreConfigSchema),
 );
 export type CommandBase = z.infer<typeof commandBaseSchema>;
